refactor: remove testing to test file
