+++ crates/committer/src/patricia_merkle_tree/test_utils.rs
@@ -1,11 +1,25 @@
 use crate::felt::Felt;
+use crate::patricia_merkle_tree::errors::TypesError;
+use crate::patricia_merkle_tree::node_data::inner_node::{EdgePathLength, PathToBottom};
+use crate::patricia_merkle_tree::node_data::leaf::SkeletonLeaf;
 use ethnum::U256;
 use rand::rngs::ThreadRng;
 use rand::Rng;
 use rstest::{fixture, rstest};
 
-use crate::patricia_merkle_tree::node_data::inner_node::{EdgePathLength, PathToBottom};
-use crate::patricia_merkle_tree::node_data::leaf::SkeletonLeaf;
+impl TryFrom<&U256> for Felt {
+    type Error = TypesError<U256>;
+    fn try_from(value: &U256) -> Result<Self, Self::Error> {
+        if *value > U256::from(&Felt::MAX) {
+            return Err(TypesError::ConversionError {
+                from: *value,
+                to: "Felt",
+                reason: "value is bigger than felt::max",
+            });
+        }
+        Ok(Self::from_bytes_be(&value.to_be_bytes()))
+    }
+}
 
 impl From<u8> for SkeletonLeaf {
     fn from(value: u8) -> Self {
